+++ lib/multiply_els.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+def multiply_els(arr)
+  arr.my_inject(1, :*)
+end
